@@ -4,14 +4,14 @@ import React from 'react';
 export default class Bar extends React.Component {
 
   static propTypes = {
-    values: React.PropTypes.any.isRequired,
-    delimiter: React.PropTypes.string,
-    fill: React.PropTypes.string,
-    height: React.PropTypes.number,
-    min: React.PropTypes.number,
-    max: React.PropTypes.number,
-    padding: React.PropTypes.number,
-    width: React.PropTypes.number,
+    values: PropTypes.any.isRequired,
+    delimiter: PropTypes.string,
+    fill: PropTypes.string,
+    height: PropTypes.number,
+    min: PropTypes.number,
+    max: PropTypes.number,
+    padding: PropTypes.number,
+    width: PropTypes.number,
   }
 
   static defaultProps = {
